初始化

+++ story-admin/src/main/java/com/story/storyadmin/domain/entity/sysmgr/Attachment.java
@@ -1,12 +1,13 @@
 package com.story.storyadmin.domain.entity.sysmgr;
 
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.story.storyadmin.domain.entity.BaseEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
-import java.io.Serializable;
 import java.util.Date;
 
 /**
@@ -41,9 +42,9 @@ public class Attachment extends BaseEntity {
     private Date uploadDate;
 
     /**
-     * 上传人
+     * 上传人Id
      */
-    private Integer uploader;
+    private Long uploader;
 
     /**
      * 1 正常 0 已被删除
@@ -51,9 +52,9 @@ public class Attachment extends BaseEntity {
     private Short status;
 
     /**
-     * 删除人
+     * 删除人Id
      */
-    private Integer deleter;
+    private Long deleter;
 
     /**
      * 删除时间
